stable

+++ app/Http/Controllers/ExtensionController.php
@@ -47,39 +47,11 @@ class ExtensionController extends Controller
      */
     public function store(StoreExtensionRequest $request)
     {
-      $password = mt_rand(100000000000, 999999999999) . '';
+      $data = $request->only($request->only( Extension::getStoreAttributes() ));
+      $data = array_merge( $data, ['admin_id'=>auth()->id()] );
+      $extension = Extension::create($data);
 
-      $extension = Extension::create([
-        'admin_id'         => auth()->user()->id,
-        'name'             => $request->name,
-
-        'phone_1'          => $request->phone_1,
-        'phone_2'          => $request->phone_2,
-        'phone_3'          => $request->phone_3,
-        'phone_4'          => $request->phone_4,
-
-        'password'         => bcrypt( $password ),
-        'email'            => $request->email,
-        'owner_phone'      => $request->owner_phone,
-        'owner_name'       => $request->owner_name,
-        'emergency_contact'=> $request->emergency_contact,
-        'emergency_contact_name'=> $request->emergency_contact_name,
-        
-        'pets_count'       => $request->pets_count ?: 0,
-        'deposit'          => $request->deposit,
-        
-        'has_own_parking'  => $request->has_own_parking,
-        'parking_number1'  => $request->parking_number1,
-        'parking_number2'  => $request->parking_number2,
-        
-        'observation'      => $request->observation,
-        'resident_id'      => $request->resident_id,
-        'resident_id_2'    => $request->resident_id_2,
-        'resident_id_3'    => $request->resident_id_3,
-        'resident_id_4'    => $request->resident_id_4,
-      ]);
-
-      if( auth()->user()->device_building_id ){
+      if( auth()->user()->device_community_id && auth()->user()->device_building_id ){
         $devices = new Devices();
         if( !$devices->addRoom($extension) ){
           $extension->delete();
